feat: add official Anthropic SDK as alternative API backend
Extract AnthropicClient protocol from runner.py with two implementations:
- HttpxAnthropicClient: preserves existing direct httpx calls with custom
  retry logic, SSE parsing, and header management
- SdkAnthropicClient: wraps anthropic.AsyncAnthropic with native retries,
  correct Python User-Agent, and proper auth handling

Config: NOUS_API_BACKEND env var (default "sdk", or "httpx" for old path).
Runner delegates _call_api/_call_api_stream to pluggable client.

Also fixes pre-existing stale test mocks (3-tuple -> 4-tuple for _tool_loop).
